Refactor code to use new Session
This introduces a `Session` type that has a life
cycle of several possible Connect. It is
now the facade to use for connecting and
publishing.

A a consequence, the earlier `connect.go`,
and `publish.go` files are no longer needed.
The busn logic for doing a publish is now
very simple and lives in the `pub` command.